Refine player visibility and last known position logic
Improved the handling of the player's last known position in BrainBehaviour and PlayerTargetSensor. Now, the last known position is only updated while the player is visible, and is frozen when the player leaves sight.
However the position is currently not updated fast enough so the agent will check an "outdated" position.