Updated ClearLastKnownAction + Brain
There is now variables in the inspector of the brain behaviour on the guard that decide the "knowledge" of the players location after pursuit ends and ClearLastKnown starts.
Can probably be adjusted with DDA so the Agent can get "smarter" and have better knowledge of your location or "dumber" and not known your location.