A lot of Goap diagnostics, made failsafe Goap Persistence
GOAPConfig got destroyed on load when playing in Build.
After a lot of debugging, the culprit was found: There was a dual instance of "Game Over Manager" one on GOAPConfig and one on a manager.
Now it has been removed from GOAPConfig and GOAP persistence has been added as a failsafe.